requirements.txt
beautifulsoup4==4.12.2
html5lib==1.1
lxml==4.9.3

requests==2.31.0

fake-useragent==1.3.0

pandas==2.1.1

selenium==4.14.0
webdriver_manager==4.0.1
textblob==0.17.1
selenium-wire==5.1.0

python-dotenv==1.0.0